requirements.txt
django
djangorestframework
psycopg2-binary
django-debug-toolbar
pep8
pylint
unidecode
pillow
django-ckeditor